修正5

+++ db/migrate/20231008065622_create_entry_users.rb
@@ -2,9 +2,9 @@ class CreateEntryUsers < ActiveRecord::Migration[6.1]
   def change
     create_table :entry_users do |t|
 
-      t.references :user, type: :integer, foreign_key: true
+      t.references :user, type: :bigint, foreign_key: true
 
-      t.references :commumity, type: :integer, foreign_key: true
+      t.references :commumity, type: :bigint, foreign_key: true
 
       t.timestamps
     end
